refactor(anki-converter): deduplicate zip metadata helper
